Move IPC stuff around
